Fix review comments: refactor dashboard routes and update tests
- Split dashboard.py into modular structure (overview.py, sync.py, graph.py)
- Move all DB queries to storage layer (dashboard_queries.py)
- Add timezone=True to DateTime columns in ORM models
- Use parameterized Cypher queries instead of f-strings
- Fix store_query_trace to properly serialize JSON
- Remove double commit in connections.py
- Use datetime.now(UTC) instead of deprecated datetime.utcnow()
- Replace time.time() with time.perf_counter() for timing
- Add workspace validation via FastAPI dependency
- Update all 28 dashboard API tests with correct mocks and fixtures
- Use directed Cypher relationships for better performance
